Support #38594 - Replace react dropzone uploader with react dropzone
- Created types to mimic the react-dropzone IFileWithMeta and IMeta.
- Updated FileUploader component to use the new dropzone component.
