+++ packages/common-ui/lib/file-dropzone/FileTypes.ts
@@ -1,0 +1,15 @@
+export interface IFileWithMeta {
+  file: File;
+  meta: IMeta;
+  cancel: () => void;
+  restart: () => void;
+  remove: () => void;
+  xhr?: XMLHttpRequest;
+}
+
+export interface IMeta {
+  type: string;
+  name: string;
+  size: number;
+  lastModifiedDate: string;
+}
