+++ packages/common-ui/lib/index.ts
@@ -146,3 +146,4 @@ export * from "./export/exportUtils";
 export * from "./export/useBlobLoad";
 export * from "./classification/TaxonomyTree";
 export * from "./file-dropzone/FileDropzone";
+export * from "./file-dropzone/FileTypes";
